+++ demos/hello-layers/src/ace/framework/layer_2.py
@@ -1,15 +1,15 @@
-from resource.layer import Layer, LayerSettings
+from ace.framework.layer import Layer, LayerSettings
 
 # TODO: Add
 PRIMARY_DIRECTIVE = ""
 
 
-class Layer1(Layer):
+class Layer2(Layer):
 
     @property
     def settings(self):
         return LayerSettings(
-            name="layer_1",
-            label="Aspirational",
+            name="layer_2",
+            label="Global Strategy",
             primary_directive=PRIMARY_DIRECTIVE
         )
