+++ demos/hello-layers/src/ace/framework/layer_5.py
@@ -1,15 +1,15 @@
-from resource.layer import Layer, LayerSettings
+from ace.framework.layer import Layer, LayerSettings
 
 # TODO: Add
 PRIMARY_DIRECTIVE = ""
 
 
-class Layer4(Layer):
+class Layer5(Layer):
 
     @property
     def settings(self):
         return LayerSettings(
-            name="layer_4",
-            label="Executive Function",
+            name="layer_5",
+            label="Cognitive Control",
             primary_directive=PRIMARY_DIRECTIVE
         )
